Trait libra_secure_storage::CryptoStorage

pub trait CryptoStorage: Send + Sync {
    fn create_key(&mut self, name: &str) -> Result<Ed25519PublicKey, Error>;

    fn export_private_key(&self, name: &str) -> Result<Ed25519PrivateKey, Error>;

    fn export_private_key_for_version(
        &self, 
        name: &str, 
        version: Ed25519PublicKey
    ) -> Result<Ed25519PrivateKey, Error>;

    fn get_public_key(&self, name: &str) -> Result<PublicKeyResponse, Error>;

    fn rotate_key(&mut self, name: &str) -> Result<Ed25519PublicKey, Error>;

    fn sign_message(
        &mut self, 
        name: &str, 
        message: &HashValue
    ) -> Result<Ed25519Signature, Error>;

    fn sign_message_using_version(
        &mut self, 
        name: &str, 
        version: Ed25519PublicKey, 
        message: &HashValue
    ) -> Result<Ed25519Signature, Error>;

    fn import_private_key(
        &mut self, 
        _name: &str, 
        _key: Ed25519PrivateKey
    ) -> Result<(), Error> { ... }
}

CryptoStorage provides an abstraction for secure generation and handling of cryptographic keys.

Required methods

fn create_key(&mut self, name: &str) -> Result<Ed25519PublicKey, Error>

Securely generates a new named Ed25519 private key. The behavior for calling this interface multiple times with the same name is implementation specific.

fn export_private_key(&self, name: &str) -> Result<Ed25519PrivateKey, Error>

Returns the Ed25519 private key stored at 'name'.

fn export_private_key_for_version(
&self,
name: &str,
version: Ed25519PublicKey
) -> Result<Ed25519PrivateKey, Error>

Returns the Ed25519 private key stored at 'name' and identified by 'version', which is the corresponding public key. This may fail even if the 'named' key exists but the version is not present.

fn get_public_key(&self, name: &str) -> Result<PublicKeyResponse, Error>

Returns the Ed25519 public key stored at 'name'.

fn rotate_key(&mut self, name: &str) -> Result<Ed25519PublicKey, Error>

Rotates an Ed25519 private key. Future calls without version to this 'named' key will return the rotated key instance. The previous key is retained and can be accessed via the version. At most two versions are expected to be retained.

fn sign_message(
&mut self,
name: &str,
message: &HashValue
) -> Result<Ed25519Signature, Error>

Signs the provided message using the 'named' private key.

fn sign_message_using_version(
&mut self,
name: &str,
version: Ed25519PublicKey,
message: &HashValue
) -> Result<Ed25519Signature, Error>

Signs the provided message using the 'named' and 'versioned' private key. This may fail even if the 'named' key exists but the version is not present.

Loading content...
Provided methods

fn import_private_key(
&mut self,
_name: &str,
_key: Ed25519PrivateKey
) -> Result<(), Error>

An optional API that allows importing private keys and storing them at the provided name. This is not intended to be used in production and the API may throw unimplemented if not used correctly. As this is purely a testing API, there is no defined behavior for importing a key for a given name if that name already exists. It only exists to allow Libra to be run in test environments where a set of deterministic keys must be generated.

Loading content...