Trait libra_secure_storage::KVStorage

pub trait KVStorage: Send + Sync {
    fn available(&self) -> bool;

    fn get(&self, key: &str) -> Result<GetResponse, Error>;

    fn set(&mut self, key: &str, value: Value) -> Result<(), Error>;

    fn reset_and_clear(&mut self) -> Result<(), Error>;
}

A secure key/value storage engine. Create takes a policy that is enforced internally by the actual backend. The policy contains public identities that the backend can translate into a unique and private token for another service. Hence get and set internally will pass the current service private token to the backend to gain its permissions.

Required methods

fn available(&self) -> bool

Returns true if the backend service is online and available.

fn get(&self, key: &str) -> Result<GetResponse, Error>

Retrieves a value from storage and fails if the backend is unavailable or the process has invalid permissions.

fn set(&mut self, key: &str, value: Value) -> Result<(), Error>

Sets a value in storage and fails if the backend is unavailable or the process has invalid permissions.

fn reset_and_clear(&mut self) -> Result<(), Error>

Resets and clears all data held in the storage engine. Note: this should only be exposed and used for testing. Resetting the storage engine is not something that should be supported in production.

Loading content...