Trait libra_secure_storage::CryptoStorage

pub trait CryptoStorage: Send + Sync {
    fn create_key(&mut self, name: &str) -> Result<Ed25519PublicKey, Error>;

    fn export_private_key(&self, name: &str) -> Result<Ed25519PrivateKey, Error>;

    fn export_private_key_for_version(
        &self, 
        name: &str, 
        version: Ed25519PublicKey
    ) -> Result<Ed25519PrivateKey, Error>;

    fn get_public_key(&self, name: &str) -> Result<PublicKeyResponse, Error>;

    fn rotate_key(&mut self, name: &str) -> Result<Ed25519PublicKey, Error>;

    fn sign_message(
        &mut self, 
        name: &str, 
        message: &HashValue
    ) -> Result<Ed25519Signature, Error>;

    fn sign_message_using_version(
        &mut self, 
        name: &str, 
        version: Ed25519PublicKey, 
        message: &HashValue
    ) -> Result<Ed25519Signature, Error>;
}

CryptoStorage offers a secure storage engine for generating, using and managing cryptographic keys securely. The API offered here is inspired by the 'Transit Secret Engine' provided by Vault: a production-ready storage engine for cloud infrastructures (e.g., see https://www.vaultproject.io/docs/secrets/transit/index.html).

Required methods

fn create_key(&mut self, name: &str) -> Result<Ed25519PublicKey, Error>

Securely generates a new named Ed25519 key pair and returns the corresponding public key. There are no guarantees about the state of the system after calling this multiple times. The behavior is implementation specific.

The new key pair is named according to the 'name'. To access or use the key pair (e.g., sign or encrypt data), subsequent API calls must refer to the key pair by name. As this API call may fail (e.g., if a key pair with the given name already exists), an error may also be returned.

fn export_private_key(&self, name: &str) -> Result<Ed25519PrivateKey, Error>

Returns the private key for a given Ed25519 key pair, as identified by the 'name'. If the key pair doesn't exist, or the caller doesn't have the appropriate permissions to retrieve the private key, this call will fail with an error.

fn export_private_key_for_version(
&self,
name: &str,
version: Ed25519PublicKey
) -> Result<Ed25519PrivateKey, Error>

Returns the private key for a given Ed25519 key pair version, as identified by the 'name' and 'version'. If the key pair at the specified version doesn't exist, or the caller doesn't have the appropriate permissions to retrieve the private key, this call will fail with an error.

fn get_public_key(&self, name: &str) -> Result<PublicKeyResponse, Error>

Returns the public key for a given Ed25519 key pair, as identified by the 'name'. If the key pair doesn't exist, or the caller doesn't have the appropriate permissions to retrieve the public key, this call will fail with an error.

fn rotate_key(&mut self, name: &str) -> Result<Ed25519PublicKey, Error>

Rotates an Ed25519 key pair by generating a new Ed25519 key pair, and updating the 'name' to reference the freshly generated key. The previous key pair is retained in storage if needed. If multiple key rotations occur over the lifetime of a key pair, only two versions of the key pair are maintained (i.e., the current and previous one). If the key pair doesn't exist, or the caller doesn't have the appropriate permissions to retrieve the public key, this call will fail with an error. Otherwise, the new public key for the rotated key pair is returned.

fn sign_message(
&mut self,
name: &str,
message: &HashValue
) -> Result<Ed25519Signature, Error>

Signs the given message using the private key associated with the given 'name'. If the key pair doesn't exist, or the caller doesn't have the appropriate permissions to retrieve and use the public key, this call will fail with an error.

fn sign_message_using_version(
&mut self,
name: &str,
version: Ed25519PublicKey,
message: &HashValue
) -> Result<Ed25519Signature, Error>

Signs the given message using the private key associated with the given 'name' and 'version'. If the key pair doesn't exist, or the caller doesn't have the appropriate permissions to perform the operation, this call will fail with an error. Note: the 'version' is specified using the public key associated with a key pair. Only two versions of a key pair are ever maintained at any given time (i.e., the current version, and the previous version).

Loading content...